pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2md"
version = "0.1.0"
description = "Convert PDFs to nicely formatted markdown files with standalone or AI-powered modes"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "pdf2md", email = "nope@example.com"},
]
keywords = ["pdf", "markdown", "converter", "cli", "gemini", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pymupdf>=1.23.0",
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
pdf2md = "pdf2md:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=22.0", "ruff>=0.1.0"]

[tool.setuptools]
py-modules = ["pdf2md"]

